FIx handling of OPc in HSS
We cannot use '==' to compare two strings in Java.  This would only
compare the references to the strings, but not the actual content.

As the database-read OP/OPc strings are dynamically allocated, their
reference (address) cannot be identical to the property.  We have to
use the .equals() method to compare the actual contents.